raise warning levels.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@3069 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/codecs/mod_ilbc/mod_ilbc.c
@@ -49,7 +49,7 @@ static switch_status_t switch_ilbc_init(switch_codec_t *codec, switch_codec_flag
 {
 	struct ilbc_context *context;
 	int encoding, decoding;
-	uint8_t ms = codec->implementation->microseconds_per_frame / 1000;
+	uint8_t ms = (uint8_t)(codec->implementation->microseconds_per_frame / 1000);
 
 
 	if (ms != 20 && ms != 30) {
